% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-smd-data.R
\name{clean_smd_data}
\alias{clean_smd_data}
\title{Load SMD data}
\usage{
clean_smd_data(smd, root_dir = ".")
}
\arguments{
\item{smd}{smd data frame containing columns DoY, Year and ts.}

\item{root_dir}{Path to the root directory.}
}
\value{
smd data_frame containing raw data and calendar variables
}
\description{
Cleans smd data for GEFCOM-D. Removes DST hours and adjusts DoY values in
leap years.
}
\details{
The DoY adjustment for leap years sets 29th Feb to have the same DoY as 28th
Feb (59) and subtracts 1 from DoY for all days after 29th Feb. This results
in both leap years and non-leap years having DoY values of 1,2,...,365.
}
\author{
Cameron Roach
}

