% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbl-block-bs.R
\name{dbl_block_bs}
\alias{dbl_block_bs}
\title{Double seasonal block bootstrap}
\usage{
dbl_block_bs(dates, start_date = today(), end_date = today() + 30,
  n_sims = 100, avg_block_len = 14, delta_loc = 3, delta_len = 3)
}
\arguments{
\item{start_date}{start date for period we want bootstrap samples for.}

\item{end_date}{end date for period we want bootstrap samples for.}

\item{n_sims}{number of simulations. Defaults to 100 simulations.}

\item{avg_block_len}{average length of blocks. Defaults to 14 days.}

\item{delta_loc}{amount to randomise current location by when sampling blocks. See details.}

\item{delta_len}{amount to randomise block length by. See details.

delta_loc shifts the current location in the simulated year by a random amount before sampling from a random historical year.}

\item{date_series}{array of dates.}
}
\value{
A data frame with two columns. One for simulation number and another for bootstrapped dates.
}
\description{
Generates bootstrapped dates using a double seasonal block bootstrap approach.
}
\author{
Cameron Roach
}

