% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{fit_models}
\alias{fit_models}
\title{Fit xgboost and vanilla models}
\usage{
fit_models(train_df, zones)
}
\arguments{
\item{train_df}{xgb train data frame.}

\item{zones}{vector of zones to fit models for.}
}
\value{
A list of fitted models.
}
\description{
Fit xgboost and vanilla models
}
