% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical-reconciliation.R
\name{format_rec_df}
\alias{format_rec_df}
\title{Format reconciled dataframes}
\usage{
format_rec_df(hts_raw, hts_rec, variable_cols, pred_col_name = "Prediction")
}
\arguments{
\item{hts_raw}{the raw time-series data.}

\item{hts_rec}{the reconciled data frame.}

\item{variable_cols}{variable column. Columns that provide a unique identifier for each time-series forecast. Inherited from hts_reconciliation().}

\item{pred_col_name}{the name assigned to the prediction column. Defaults to "Prediction".}
}
\description{
This function is used after reconciliation to convert data to the correct format.
Used by hts_reconciliation function.
}
