% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-calendar-vars.R
\name{get_calendar_vars}
\alias{get_calendar_vars}
\title{Get calendar variables}
\usage{
get_calendar_vars(x, root_dir = ".")
}
\arguments{
\item{x}{data frame containing date and hour data.}
}
\value{
A data frame containing calendar variables.
}
\description{
Gets calendar variables such as holiday, hour, day of year, etc.
}
\author{
Cameron Roach
}
