% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dummy-vars.R
\name{get_dummy_vars}
\alias{get_dummy_vars}
\title{Get dummy variables}
\usage{
get_dummy_vars(x, calendar_vars = TRUE, lags = FALSE)
}
\arguments{
\item{x}{cleaned data frame. Should be the output from \code{clean_smd_data()} function.}

\item{calendar_vars}{logical. Indicates if calendar dummy variables should be returned.}

\item{lags}{logical. Indicates if lagged variables should be returned.}
}
\value{
A data frame containing extra dummy variables.
}
\description{
Gets dummy variables for GEFCOM smd data. Dummy variables include calendar
variables and lagged dry bulb and dew point temperature variables.
}
\author{
Cameron Roach
}

