% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lagged-vars.R
\name{get_lagged_vars}
\alias{get_lagged_vars}
\title{Get lagged variables}
\usage{
get_lagged_vars(x)
}
\arguments{
\item{x}{cleaned data frame. Should be the output from \code{clean_smd_data()} function.}
}
\value{
A data frame containing lagged variables.
}
\description{
Calculates lagged temperature variables for GEFCOM smd data.
}
\author{
Cameron Roach
}

