% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_quantiles}
\alias{get_quantiles}
\title{Get quantiles}
\usage{
get_quantiles(x, prediction_col)
}
\arguments{
\item{x}{a data frame containing simulated demand.}

\item{prediction_col}{string for prediction column name.}
}
\value{
A data frame containing quantile forecasts.
}
\description{
Get quantiles
}
