% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rmse}
\alias{get_rmse}
\title{Calculate RMSE}
\usage{
get_rmse(x)
}
\arguments{
\item{x}{data frame with Demand and Prediction columns}
}
\value{
data frame with RMSE values
}
\description{
Calculate RMSE
}
