% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical-reconciliation.R
\name{h_rec}
\alias{h_rec}
\title{Hierarchical reconciliation by OLS}
\usage{
h_rec(x, S)
}
\arguments{
\item{x}{vector of time-series forecasts for a given time t.}

\item{S}{summing matrix.}
}
\description{
Hierarchical reconciliation by OLS
}
