% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical-reconciliation.R
\name{h_rec_w}
\alias{h_rec_w}
\title{Hierarchical reconciliation by WLS}
\usage{
h_rec_w(x, S, W)
}
\arguments{
\item{x}{vector of time-series forecasts for a given time t.}

\item{S}{summing matrix.}

\item{W}{weight matrix.}
}
\description{
Hierarchical reconciliation by WLS
}
