% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical-reconciliation.R
\name{hts_reconciliation}
\alias{hts_reconciliation}
\title{Hierarchical time-series reconciliation}
\usage{
hts_reconciliation(fcsts, S, W_mean, W_median, W_res_var,
  prediction_col = "Prediction", variable_cols = c("Simulation", "ts",
  "Date", "Hour"), zone_cols = c("TOTAL", "ME", "NH", "VT", "CT", "RI",
  "MASS", "SEMASS", "WCMASS", "NEMASSBOST"))
}
\arguments{
\item{fcsts}{forecast data frame.}

\item{S}{summing matrix.}

\item{variable_cols}{variable column. Columns that provide a unique
identifier for each time-series forecast.}

\item{zone_cols}{zones of the hierarchy. Need to be in the same order as for
the summing matrix.}
}
\description{
Function to reconcile time-series forecasts. Returns reconciled forecasts
using four methods: optimal (summing matrix), mean, median and residual
variance.
}
