% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_smd_data}
\alias{load_smd_data}
\title{Load SMD data}
\usage{
load_smd_data(load_zones)
}
\arguments{
\item{load_zones}{Electricity load zones.}
}
\value{
smd data_frame containing raw data and calendar variables
}
\description{
Loads data for GEFCom2017-D.
}
\author{
Cameron Roach
}

