% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinball-loss.R
\name{pinball_loss}
\alias{pinball_loss}
\title{Pinball loss function}
\usage{
pinball_loss(tau, y, q)
}
\arguments{
\item{tau}{integer 1, 2, ... 99. Quantile to calculate pinball loss score
for.}

\item{y}{numeric. Observed value.}

\item{q}{numeric. Predicted value for quantile tau.}
}
\value{
Pinball loss score.
}
\description{
Calculates the pinball loss score for a given quantile.
}
\author{
Cameron Roach
}

