% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid-block-bs.R
\name{resid_block_bs}
\alias{resid_block_bs}
\title{Block bootstrap}
\usage{
resid_block_bs(x_resid, x_fcst, block_length = 4)
}
\arguments{
\item{x_resid}{data frame containing residuals.}

\item{x_fcst}{data frame containing forecasts and dates.}

\item{block_length}{numeric. Length of bootstrap blocks.}
}
\value{
The forecast data frame with a new residuals column.
}
\description{
Generates bootstrapped residuals using a block bootstrapping approach. Works
in a different fashion to \code{dbl_block_bs} as it must apply the
bootstrapped residuals to forecast data frame. Runs several checks to remove
incomplete days from historical residuals.
}
\details{
TODO: change this so that forecast data frame isn't input. Can just return a
vector of residuals that gets added to data frame.
}
\author{
Cameron Roach
}
