% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-similar-day.R
\name{sample_sequence_similar_days}
\alias{sample_sequence_similar_days}
\title{Sample sequence of similar days}
\usage{
sample_sequence_similar_days(fcst_dates, resid_dates, min_block_length = 6,
  max_block_length = 14, date_drift = 7)
}
\arguments{
\item{fcst_dates}{vector containing forecasts dates. These are the dates we
wish to find similar dates for.}

\item{resid_dates}{data frame of residual date info. Must contain Date
column. Only unique values should be included.}

\item{min_block_length}{integer giving minimum block length in days.}

\item{max_block_length}{integer giving maximum block length in days.}

\item{date_drift}{integer specifying number of days that block start length
can differ from current day of year.}
}
\value{
The forecast data frame with a new residuals column.
}
\description{
Takes a vector of dates as inputs and randomly samples a random number of
from the residual data frame. Similar dates can be defined by day of the week
or working/non-working days.
}
\author{
Cameron Roach
}
