% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-results.R
\name{save_results}
\alias{save_results}
\title{Output quantile forecasts to excel}
\usage{
save_results(x, file)
}
\arguments{
\item{x}{quantile forecasts}

\item{file}{excel file name.}
}
\value{
Saves an excel file with quantile forecasts for each zone.
}
\description{
Output quantile forecasts to excel
}
