% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle-weather.R
\name{shuffle_weather_xgb}
\alias{shuffle_weather_xgb}
\title{Shuffle historical weather}
\usage{
shuffle_weather_xgb(x, fcst_start_date, fcst_end_date)
}
\arguments{
\item{x}{data frame containing historical data}

\item{fcst_start_date}{start date of forecasts}

\item{fcst_end_date}{end date for forecasts}
}
\value{
Data frame containing shuffled weather scenarios.
}
\description{
Shuffle historical weather
}
