% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_model_fat}
\alias{trim_model_fat}
\title{Trims models}
\usage{
trim_model_fat(x)
}
\arguments{
\item{x}{model}
}
\description{
Trims the fat - remove list elements that are large and not needed.
}
\details{
Note: use length(serialize(x, NULL)) as a proxy for size. Better than object.size().
}
