% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{fit_models}
\alias{fit_models}
\title{Fit xgboost and vanilla models}
\usage{
fit_models(vanilla_train_df, xgb_train_df)
}
\arguments{
\item{vanilla_train_df}{training data frame for vanilla model.}

\item{xgb_train_df}{xgb train data frame.}
}
\value{
A list of fitted models.
}
\description{
Fit xgboost and vanilla models
}
