% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lagged-vars.R
\name{get_lagged_vars}
\alias{get_lagged_vars}
\title{Get lagged variables}
\usage{
get_lagged_vars(x, variables, lags = 1)
}
\arguments{
\item{x}{cleaned data frame. Should be the output from \code{clean_smd_data()} function.}

\item{variables}{character vector. Vector of variable names that should be lagged.}

\item{lags}{numeric vector. Vector of lags to calculate. Hourly values.}
}
\value{
A data frame containing lagged variables.
}
\description{
Calculates lagged temperature variables for GEFCOM smd data.
}
\author{
Cameron Roach
}

