% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-smd-data.R
\name{get_train_test_df}
\alias{get_train_test_df}
\title{Get train and test data frames}
\usage{
get_train_test_df(vanilla_train_start_date, vanilla_train_end_date,
  xgb_train_start_date, xgb_train_end_date)
}
\arguments{
\item{vanilla_train_start_date}{start date for vanilla model's training data.}

\item{vanilla_train_end_date}{end date for vanilla model's training data.}

\item{xgb_train_start_date}{start date for XGBoost model's training data.}

\item{xgb_train_end_date}{end date for XGBoost model's training data.}
}
\value{
List containing xgb train, vanilla train and test data frames.
}
\description{
Get train and test data frames
}
