% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-similar-day.R
\name{sample_sequence_similar_days}
\alias{sample_sequence_similar_days}
\title{Sample sequence of similar days}
\usage{
sample_sequence_similar_days(fcst_dates, resid_dates)
}
\arguments{
\item{fcst_dates}{vector containing forecasts dates. These are the dates we
wish to find similar dates for.}

\item{resid_dates}{data frame of residual date info. Must contain Date
column. Only unique values should be included.}
}
\value{
The forecast data frame with a new residuals column.
}
\description{
Takes a vector of dates as inputs and randomly samples a random number of 
from the residual data frame. Similar dates can be defined by day of the week
or working/non-working days.
}
\author{
Cameron Roach
}
