% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-similar-day.R
\name{sample_similar_day}
\alias{sample_similar_day}
\title{Sample similar days}
\usage{
sample_similar_day(x_resid, x_fcst)
}
\arguments{
\item{x_resid}{data frame containing residuals.}

\item{x_fcst}{data frame containing forecasts and dates.}
}
\value{
The forecast data frame with a new residuals column.
}
\description{
Takes a date as an input and randomly samples another similar date. Similar
dates can be defined by day of the week or working/non-working days. A window
that indicates how many months eitherside of the current date can be used for
sampling can also be specified.
}
\details{
TODO: filter for adjacent (+-3) months when selecting residuals - get them from roughly the same time of year
TODO: split up for holidays and non-holidays

TODO: change this so that forecast data frame isn't input. Can just return a
vector of residuals that gets added to data frame.
}
\author{
Cameron Roach
}
